SimSpin.jl Documentation

Usage

For the installation procedure of the SimSpin package please follow the installation instructions on the package's README.

Once installed, a kinematic datacube for the example galaxy observed from redshift 0.05 and inclined at 30 degrees from face on can be built using the following procedure:

> sim_data = SimSpin.sim_data("path/to/SimSpin/example/SimSpin_example.hdf5")
> datacube = SimSpin.build_datacube(sim_data, z = 0.05, inc_deg = 30)

Functions

SimSpin.sim_data — Function

sim_data(filename;
        pytpe = [],
        ssp = false)

Reads in a SimSpin format HDF5 file at location, filename. Returns array of Sim_particles.

Keyword arguments (optional):

ptype       A vector of the particles types to be read in e.g. ptype = [1,3].
            If omitted all particles types will be read.
ssp         Boolean value to use ssp particle information.

SimSpin.obs_data_prep — Function

obs_data_prep(galaxy_data;
                r200 = 200,
                z = 0.05,
                fov = 15.,
                ap_shape = "circular",
                central_wvl = 4800,
                lsf_fwhm = 2.65,
                pixel_sscale = 0.5,
                pixel_vscale = 1.04,
                inc_deg = 70,
                filter = "r",
                blur = nothing)

This function prepares the particle data for a given observation.

Returns:

galaxy_data         Array of particle data formatted for the observation specified by the parameters.
parts_in_cell       3D array of the particles corresponding to each element in the IFU data-cube.
ap_region           The aperture region mask used to remove flux outside of the specified aperture.
sbin                The number of spatial bins in the aperture.
vbin                The number of velocity bins in the flux grid.

Keyword arguments (optional):

r200            The virial radius specified in the simulation, kpc.
z               The projected redshift at which the observation is made.
fov             The field of view of the IFU, diameter in arcseconds.
ap_shape        The shape of the field of view, with options "circular", "square" or "hexagonal".
central_wvl     The central filter wavelength used for the observation, given in angstroms.
lsf_fwhm        The line spread function full-width half-max, given in angstroms.
pixel_sscale    The corresponding spatial pixel scale associated with a given telescope output in arcseconds.
pixel_vscale    The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.
inc_deg         The inclination at which to observe the galaxy in degrees.
filter          If particles type is ssp, the filter within which the SED is generated. Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.
blur            Specified to apply observational seeing effects to the cube. A list of the form "psf" = "Moffat", "fwhm" = 0.5. "psf" specifies the shape of the PSF chosen and may be either "Moffat" or "Gaussian". "fwhm" is a numeric specifying the full-width half-maximum of the PSF given in units of arcseconds.

SimSpin.build_datacube — Function

build_datacube(galaxy_data;
                r200 = 200,
                z = 0.05,
                fov = 15.,
                ap_shape = "circular",
                central_wvl = 4800,
                lsf_fwhm = 2.65,
                pixel_sscale = 0.5,
                pixel_vscale = 1.04,
                inc_deg = 70,
                filter = "r",
                blur = nothing)

Returns a simulated ifu datacube for input, galaxy_data, an array of Particle. Observational parameters specified as keyword arguments.

Keyword arguments (optional):

r200            The virial radius specified in the simulation, kpc.
z               The projected redshift at which the observation is made.
fov             The field of view of the IFU, diameter in arcseconds.
ap_shape        The shape of the field of view, with options "circular", "square" or "hexagonal".
central_wvl     The central filter wavelength used for the observation, given in angstroms.
lsf_fwhm        The line spread function full-width half-max, given in angstroms.
pixel_sscale    The corresponding spatial pixel scale associated with a given telescope output in arcseconds.
pixel_vscale    The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.
inc_deg         The inclination at which to observe the galaxy in degrees.
filter          If particles type is ssp, the filter within which the SED is generated. Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.
blur            Specified to apply observational seeing effects to the cube. A list of the form "psf" = "Moffat", "fwhm" = 0.5. "psf" specifies the shape of the PSF chosen and may be either "Moffat" or "Gaussian". "fwhm" is a numeric specifying the full-width half-maximum of the PSF given in units of arcseconds.

SimSpin.flux_grid — Function

flux_grid(parts_in_cell,
            ap_region,
            sbin,
            vbin,
            redshift,
            filter)

Computes the fluxes for each element of the IFU data-cube.

The purpose of this function is to construct the mock flux values within each cell of the IFU cube. It accepts output parameters from obs_data_prep() and returns a 3D array containing the flux at each cell position due to contributions from the particles. If ssd particles are supplied, an SED is generated in each cell using ProSpect. Else, the luminosity in each cell is converted to flux.

Parameters:

parts_in_cell       3D array of the particles corresponding to each element in the IFU data-cube.
ap_region           The aperture region mask used to remove flux outside of the specified aperture.
sbin                The number of spatial bins in the aperture.
vbin                The number of velocity bins in the flux grid.
redshift            The projected redshift at which the observation is made.
filter              If ssp particles are supplied, the filter within which the SED is generated.
                    Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.