SimSpin.jl Documentation

Usage

For the installation procedure of the SimSpin package please follow the installation instructions on the package's README.

Once installed, a kinematic datacube for the example galaxy observed from redshift 0.05 and inclined at 30 degrees from face on can be built using the following procedure:

> sim_data = SimSpin.sim_data("path/to/SimSpin/example/SimSpin_example.hdf5")
> datacube = SimSpin.build_datacube(sim_data, z = 0.05, inc_deg = 30)

Functions

SimSpin.Blur — Type

Blur(psf; sigma, fwhm)

Create a struct containing the blurring to be used.
Must specify blurring point spread function, `psf`,
currently only "Gaussian" is supported.

Keyword arguments (one must be specified, sigma is prioritised):

    sigma       The standard deviation of the point spread function
    fwhm        The full width half max of the point spread function

SimSpin.build_datacube — Function

build_datacube(galaxy_data;
                r200 = 200,
                z = 0.05,
                fov = 15.,
                ap_shape = "circular",
                central_wvl = 4800,
                lsf_fwhm = 2.65,
                pixel_sscale = 0.5,
                pixel_vscale = 1.04,
                inc_deg = 70,
                filter = "r",
                blur = Blur("none"))

Returns a simulated ifu datacube for input, galaxy_data, an array of Particle. Observational parameters specified as keyword arguments.

Keyword arguments (optional):

r200            The virial radius specified in the simulation, kpc.
z               The projected redshift at which the observation is made.
fov             The field of view of the IFU, diameter in arcseconds.
ap_shape        The shape of the field of view, with options "circular", "square" or "hexagonal".
central_wvl     The central filter wavelength used for the observation, given in angstroms.
lsf_fwhm        The line spread function full-width half-max, given in angstroms.
pixel_sscale    The corresponding spatial pixel scale associated with a given telescope output in arcseconds.
pixel_vscale    The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.
inc_deg         The inclination at which to observe the galaxy in degrees.
filter          If particles type is ssp, the filter within which the SED is generated. Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.
blur            Specified to apply observational seeing effects to the cube. Use `Blur()` function to create blur profile. If omitted no blurring occurs.

SimSpin.flux_grid — Function

flux_grid(parts_in_cell,
            ap_region,
            sbin,
            vbin,
            redshift,
            filter)

Computes the fluxes for each element of the IFU data-cube.

The purpose of this function is to construct the mock flux values within each cell of the IFU cube. It accepts output parameters from obs_data_prep() and returns a 3D array containing the flux at each cell position due to contributions from the particles. If ssd particles are supplied, an SED is generated in each cell using ProSpect. Else, the luminosity in each cell is converted to flux.

Parameters:

parts_in_cell       1D array of the particles corresponding to each element in the IFU data-cube.
ap_region           The aperture region mask used to remove flux outside of the specified aperture.
sbin                The number of spatial bins in the aperture.
vbin                The number of velocity bins in the flux grid.
redshift            The projected redshift at which the observation is made.
filter              If ssp particles are supplied, the filter within which the SED is generated.
                    Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.

SimSpin.ifu_cube — Function

ifu_cube(flux_grid,
            parts_in_cell,
            sbin,
            vbin,
            vseq,
            lsf_size)

The purpose of this function is to construct an IFU data cube. It accepts the fluxgrid output by the `fluxgrid()` function and returns a similar, IFU-like, 3D array where each particle's flux contributes a Gaussian distribution in the velocity axis.

Parameters:

flux_grid       Flux grid output by `flux_grid()`
parts_in_cell   1D array of the particles corresponding to each element in the IFU data-cube.
sbin            The number of spatial bins in the aperture.
vbin            The number of velocity bins in the flux grid.
vseq            The bounds of each velocity bin in the flux grid.
lsf_size        The Gaussian standard deviation of the line spread function in km/s.

SimSpin.obs_data_prep — Function

obs_data_prep(galaxy_data;
                r200 = 200,
                z = 0.05,
                fov = 15.,
                ap_shape = "circular",
                central_wvl = 4800,
                lsf_fwhm = 2.65,
                pixel_sscale = 0.5,
                pixel_vscale = 1.04,
                inc_deg = 70)

This function prepares the particle data for a given observation.

Returns:

galaxy_data         Array of particle data formatted for the observation specified by the parameters.
parts_in_cell       3D array of the particles corresponding to each element in the IFU data-cube.
ap_region           The aperture region mask used to remove flux outside of the specified aperture.
sbin                The number of spatial bins in the aperture.
vbin                The number of velocity bins in the flux grid.
vseq                The bounds of each velocity bin in the flux grid.
lsf_size            The Gaussian standard deviation of the line spread function in km/s.
ang_size            The angular size given redshift z in kpc
sbinsize            The spatial bin size in kpc per bin

Keyword arguments (optional):

r200            The virial radius specified in the simulation, kpc.
z               The projected redshift at which the observation is made.
fov             The field of view of the IFU, diameter in arcseconds.
ap_shape        The shape of the field of view, with options "circular", "square" or "hexagonal".
central_wvl     The central filter wavelength used for the observation, given in angstroms.
lsf_fwhm        The line spread function full-width half-max, given in angstroms.
pixel_sscale    The corresponding spatial pixel scale associated with a given telescope output in arcseconds.
pixel_vscale    The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.
inc_deg         The inclination at which to observe the galaxy in degrees.

SimSpin.sim_data — Function

sim_data(filename;
        pytpe = [],
        ssp = false)

Reads in a SimSpin format HDF5 file at location, filename. Returns array of Sim_particles.

Keyword arguments (optional):

ptype       A vector of the particles types to be read in e.g. ptype = [1,3].
            If omitted all particles types will be read.
ssp         Boolean value to use ssp particle information.

Data Input Format

Here we outline the expected file format accepted by SimSpin. If you would like to generate this file automatically, a short Python function has been written that uses the pynbody package to read in various simulation data types and generate a SimSpin compatible HDF5 file. See create_SimSpinFile.

If you would rather generate the SimSpin file independently, the expected file format is outlined below.

> SimSpin_example.hdf5

>> /PartType0           # Each particle type included in the simulation has its own group.
>>> /PartType0/Mass     # Each group then has a series of data sets assocaited,
>>> /PartType0/vx       #   including the position, velocity and Mass of each particle.
>>> /PartType0/vy
>>> /PartType0/vz
>>> /PartType0/x
>>> /PartType0/y
>>> /PartType0/z

>> /PartType1
>>> ...

We use the same PartType definition as Gadget: PartTypeX where 0 - gas, 1 - dark matter, 2 - disc, 3 - bulge, 4 - stars. For PartType0-3, each PartType group contains the same data sets as above. If the simulation contains stars, the Age and Metallicity information for each particle is also included:

> SimSpin_example.hdf5
>> /PartType4
>>> /PartType4/Age
>>> /PartType4/Mass
>>> /PartType4/Metallicity
>>> /PartType4/vx        
>>> /PartType4/vy
>>> /PartType4/vz
>>> /PartType4/x
>>> /PartType4/y
>>> /PartType4/z

If the file is set up in this way, the simulation data can easily be read into the SimSpin package.

References

A. Pontzen, R Roskar, G. Stinson and R. Woods, (2013), "pynbody: N-Body/SPH analysis for python", Astrophysics Source Code Library, record ascl:1305.002