SimSpin.jl Documentation

Usage

For the installation procedure of the SimSpin package please follow the installation instructions on the package's README.

Once installed, a simple procedure of three steps is required to take an observation and generate a datacube:

Create a telescope object. This specifies the field of view to be used, the aperture shape, etc. In this example we will use the default SAMI telescope constructor. See Telescope Constructors for other default and customisable constructors.

> telescope = SimSpin.SAMI()

Read in the simulation's particle data

> sim_data = SimSpin.sim_data("path/to/SimSpin/example/SimSpin_example.hdf5")

Build the datacube as a combination of a telescope, the galaxy particle data and some observation parameters (galaxy redshift, inclination and virial radius).

> datacube = SimSpin.build_datacube(sim_data, telescope, 0.05, 70, 200)

Functions

SimSpin.build_datacube — Function

build_datacube(galaxy_data, observation, ifu, blur)

Returns a simulated ifu datacube for input, galaxy_data, an array of Particle. Observational variables used are as specified in Observation and IFU types. Blurring can be added if desired using Blur type.

Parameters:

galaxy_data         Array of `Particle` describing galaxy
ifu                 Struct of type `Telescope`
z                   The projected redshift at which the observation is made.
inc_deg             The inclination at which to observe the galaxy in degrees.
r200                The virial radius specified in the simulation, kpc.
blur                Optional. Struct of type `Blur`. If omitted no blurring occurs.

SimSpin.flux_grid — Function

flux_grid(parts_in_cell,
            observe,
            filter)

Computes the fluxes for each element of the IFU data-cube.

The purpose of this function is to construct the mock flux values within each cell of the IFU cube. It accepts output parameters from obs_data_prep() and returns a 3D array containing the flux at each cell position due to contributions from the particles. If ssd particles are supplied, an SED is generated in each cell using ProSpect. Else, the luminosity in each cell is converted to flux.

Parameters:

parts_in_cell       1D array of the particles corresponding to each element in the IFU data-cube.
observe             Struct of type `Observation` containing all observation parameters.
filter              If ssp particles are supplied, the filter within which the SED is generated.
                    Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.

SimSpin.ifu_cube — Function

ifu_cube(flux_grid,
            parts_in_cell,
            observe)

The purpose of this function is to construct an IFU data cube. It accepts a flux grid in the format output by the flux_grid() function and returns a similar, IFU-like, 3D array where each particle's flux contributes a Gaussian distribution in the velocity axis.

Parameters:

flux_grid       Flux grid output by `flux_grid()`
parts_in_cell   1D array of the particles corresponding to each element in the IFU data-cube.
observe         Struct of type `Observation` containing all observation parameters.

SimSpin.obs_data_prep — Function

obs_data_prep(galaxy_data, observation, ifu)

This function prepares the particle data for a given observation with the given telescope.

Returns:

galaxy_data         Array of particle data formatted for the observation specified by the parameters.
parts_in_cell       3D array of the particles corresponding to each element in the IFU data-cube.
observed            Struct of type `Observation` containing all observation parameters.

Parameters:

galaxy_data         Array of `Particle` describing galaxy
ifu                 Struct of type `Telescope`
z                   The projected redshift at which the observation is made.
inc_deg             The inclination at which to observe the galaxy in degrees.
r200                The virial radius specified in the simulation, kpc.

SimSpin.sim_data — Function

sim_data(filename;
        pytpe = [],
        ssp = false)

Reads in a SimSpin format HDF5 file at location, filename. Returns array of Sim_particles.

Keyword arguments (optional):

ptype       A vector of the particles types to be read in e.g. ptype = [1,3].
            If omitted all particles types will be read.
ssp         Boolean value to use ssp particle information.

Constructors

Telescope Constructors

An IFU object denotes all the parameters required to make a mock observation. Any generic IFU can be made using the IFU() function. Default constructors can also be used to emulate famous IFU survey instruments. Currently SAMI, MaNGA, CALIFA and Hector are supported.

SimSpin.IFU — Type

IFU(fov, ap_shape, pixel_sscale, pixel_vscale, central_wvl, lsf_fwhm, filter)

Creates a customisable, mock IFU telescope which can be used to "observe" simulations.

Parameters:

fov             The field of view of the IFU, diameter in arcseconds.
ap_shape        The shape of the field of view, with options "circular", "square" or "hexagonal".
central_wvl     The central filter wavelength used for the observation, given in angstroms.
lsf_fwhm        The line spread function full-width half-max, given in angstroms.
pixel_sscale    The corresponding spatial pixel scale associated with a given telescope output in arcseconds.
pixel_vscale    The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.
filter          Optional. If particles type is ssp, the filter within which the SED is generated. Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.

SimSpin.SAMI — Function

SAMI()

Creates an IFU using parameters of the SAMI survey.

SimSpin.MaNGA — Function

MaNGA()

Creates an IFU using parameters of the MaNGA survey.

SimSpin.CALIFA — Function

CALIFA()

Creates an IFU using parameters of the CALIFA survey.

SimSpin.Hector — Function

Hector()

Creates an IFU using parameters of the CALIFA survey.

Blur Constructors

SimSpin.Gaussian_blur — Type

Gaussian_blur(;sigma, fwhm)

Create a struct containing seeing information.

Keyword arguments (at least one must be specified, sigma is prioritised):

sigma       The standard deviation of the point spread function
fwhm        The full width half max of the point spread function

SimSpin.Moffat_blur — Type

Moffat_blur(β;
            α,
            fwhm)

Create a struct containing seeing information. β and either α or fwhm must be specified. If both α and fwhm are specified, α is prioritised.

Arguments:

β           The power component in the Moffat distribution
α           The core width of the Moffat distribution (optional)
fwhm        The full width half max of the Moffat distribution (optional)

Dark Matter Constructors

SimSpin.Hernquist — Type

Hernquist(;dm_mass = 186.9,
            dm_a = 34.5)

Creates a Hernquist analytic dark matter mass potential. Default values are shown above.

SimSpin.NFW — Type

NFW(;dm_vm = 186.9,
        dm_a = 34.5,
        dm_rhof = 0.035)

Creates an NFW analytic dark matter mass potential. Default values are shown above.

Multi-Threading

The SimSpin package has multi-threading enabled in some critical functions. To use SimSpin with x threads (where x is the integer number of threads desired) you must call

export JULIA_NUM_THREADS=x

in a Terminal before the Julia REPL is started. This environment variable defaults to 1 if not set before the session has begun.

Data Input Format

Here we outline the expected file format accepted by SimSpin. If you would like to generate this file automatically, a short Python function has been written that uses the pynbody package to read in various simulation data types and generate a SimSpin compatible HDF5 file. See create_SimSpinFile.

If you would rather generate the SimSpin file independently, the expected file format is outlined below.

> SimSpin_example.hdf5

>> /PartType0           # Each particle type included in the simulation has its own group.
>>> /PartType0/Mass     # Each group then has a series of data sets assocaited,
>>> /PartType0/vx       #   including the position, velocity and Mass of each particle.
>>> /PartType0/vy
>>> /PartType0/vz
>>> /PartType0/x
>>> /PartType0/y
>>> /PartType0/z

>> /PartType1
>>> ...

We use the same PartType definition as Gadget: PartTypeX where 0 - gas, 1 - dark matter, 2 - disc, 3 - bulge, 4 - stars. For PartType0-3, each PartType group contains the same data sets as above. If the simulation contains stars, the Age and Metallicity information for each particle is also included:

> SimSpin_example.hdf5
>> /PartType4
>>> /PartType4/Age
>>> /PartType4/Mass
>>> /PartType4/Metallicity
>>> /PartType4/vx        
>>> /PartType4/vy
>>> /PartType4/vz
>>> /PartType4/x
>>> /PartType4/y
>>> /PartType4/z

If the file is set up in this way, the simulation data can easily be read into the SimSpin package.

References

A. Pontzen, R Roskar, G. Stinson and R. Woods, (2013), "pynbody: N-Body/SPH analysis for python", Astrophysics Source Code Library, record ascl:1305.002